+++ app/src/main/java/io/fastpix/app/VideoListScreen.kt
@@ -2,13 +2,10 @@ package io.fastpix.app
 
 import android.content.Intent
 import android.os.Bundle
-import androidx.annotation.OptIn
 import androidx.appcompat.app.AppCompatActivity
 import androidx.media3.common.util.UnstableApi
 import androidx.recyclerview.widget.LinearLayoutManager
 import io.fastpix.app.databinding.ActivityVideoListScreenBinding
-import java.util.UUID
-import kotlin.jvm.java
 
 class VideoListScreen : AppCompatActivity() {
     private lateinit var binding: ActivityVideoListScreenBinding
@@ -16,7 +13,7 @@ class VideoListScreen : AppCompatActivity() {
         VideoAdapter()
     }
 
-    @OptIn(UnstableApi::class)
+    @UnstableApi
     override fun onCreate(savedInstanceState: Bundle?) {
         super.onCreate(savedInstanceState)
         binding = ActivityVideoListScreenBinding.inflate(layoutInflater)
@@ -27,12 +24,17 @@ class VideoListScreen : AppCompatActivity() {
 
         videoAdapter.onVideoClick = { video ->
             val intent = Intent(this, MainActivity::class.java)
+            intent.putExtra(VIDEO_MODEL, video)
+            intent.putExtra(AUTO_PLAY, binding.sAutoPlay.isChecked)
+            intent.putExtra(LOOP, binding.sLoop.isChecked)
             startActivity(intent)
         }
     }
 
 
     companion object {
-        val viewerId = UUID.randomUUID().toString()
+        const val VIDEO_MODEL = "video_model"
+        const val AUTO_PLAY = "auto_play"
+        const val LOOP = "loop"
     }
 }